# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "egras API",
    "description": "Enterprise-ready Rust application seed — tenants, security & audit",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/echo": {
      "get": {
        "tags": [
          "echo"
        ],
        "operationId": "get_echo",
        "responses": {
          "200": {
            "description": "Echo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EchoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.errors.ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.errors.ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "echo"
        ],
        "operationId": "post_echo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Echo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EchoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.errors.ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.errors.ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/features": {
      "get": {
        "tags": [
          "features"
        ],
        "operationId": "get_definitions",
        "responses": {
          "200": {
            "description": "Feature catalog (operator only)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureDefinition"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.errors.ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.errors.ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "security"
        ],
        "operationId": "get_list_users",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "org_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cursor or limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "EchoResponse": {
        "type": "object",
        "description": "Response body returned by both GET and POST /api/v1/echo.",
        "required": [
          "method",
          "payload",
          "org_id",
          "principal_user_id",
          "received_at"
        ],
        "properties": {
          "key_id": {
            "type": "string",
            "format": "uuid",
            "description": "API key ID when authenticated via API key; `null` for JWT callers.",
            "nullable": true
          },
          "method": {
            "type": "string",
            "description": "HTTP method of the incoming request (\"GET\" or \"POST\")."
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organisation the caller belongs to."
          },
          "payload": {
            "type": "object",
            "description": "The JSON payload sent in the request body, or `null` for GET.",
            "nullable": true
          },
          "principal_user_id": {
            "type": "string",
            "format": "uuid",
            "description": "The principal user ID (human user or service account user)."
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "description": "Server-side timestamp at which the request was received (RFC 3339)."
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "RFC 7807 problem body returned on all error responses.\n\nAll six stable fields are present in every response; `errors` is included\nonly on validation errors (HTTP 400) and maps field name → list of slugs.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "errors": {
            "type": "object",
            "description": "Field-level validation errors (present only on 400 responses).",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "instance": {
            "type": "string",
            "description": "URI reference identifying the specific occurrence of the problem.",
            "nullable": true
          },
          "request_id": {
            "type": "string",
            "description": "Correlation ID for request tracing.",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary of the error."
          },
          "type": {
            "type": "string",
            "description": "A URI reference identifying the error type."
          }
        }
      },
...
      "FeatureDefinition": {
        "type": "object",
        "required": [
          "slug",
          "value_type",
          "default_value",
          "description",
          "self_service"
        ],
        "properties": {
          "default_value": {
            "type": "object"
          },
          "description": {
            "type": "string"
          },
          "self_service": {
            "type": "boolean"
          },
          "slug": {
            "type": "string"
          },
          "value_type": {
            "$ref": "#/components/schemas/FeatureValueType"
          }
        }
      },
...
      "FeatureValueType": {
        "type": "string",
        "enum": [
          "bool",
          "string",
          "int",
          "enum_set",
          "json"
        ]
      },
...
      "ListUsersResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSummaryDto"
            }
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MembershipDto": {
        "type": "object",
        "required": [
          "org_id",
          "org_name",
          "role_codes",
          "joined_at"
        ],
        "properties": {
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "org_name": {
            "type": "string"
          },
          "role_codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UserSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "created_at",
          "memberships"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MembershipDto"
            }
          },
          "username": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}